+++ libs/components/src/lib/components/overlay/components/bottom-sheet/constants/bottom-sheet.constants.ts
@@ -5,16 +5,6 @@ import { BottomSheetConfig } from '../types';
 export const BOTTOM_SHEET_MIN_SWIPE_TO_CLOSE_LENGTH = 150;
 export const BOTTOM_SHEET_MIN_VELOCITY_TO_CLOSE = 150;
 
-export const BOTTOM_SHEET_TRANSITION_DURATION_PROPERTY = '--et-bottom-sheet-transition-duration';
-
-export const BOTTOM_SHEET_ANIMATION_CLASSES = {
-  opening: 'et-bottom-sheet--opening',
-  open: 'et-bottom-sheet--open',
-  closing: 'et-bottom-sheet--closing',
-  closed: 'et-bottom-sheet--closed',
-};
-
-/** Injection token that can be used to access the data that was passed in to a bottom sheet. */
 export const BOTTOM_SHEET_DATA = new InjectionToken<unknown>('BottomSheetData');
 
 export const BOTTOM_SHEET_DEFAULT_OPTIONS = new InjectionToken<BottomSheetConfig>('BottomSheetDefaultOptions');
@@ -42,6 +32,4 @@ export const BOTTOM_SHEET_DEFAULT_CONFIG: BottomSheetConfig = {
   autoFocus: 'dialog',
   customAnimated: false,
   restoreFocus: true,
-  enterAnimationDuration: 300,
-  exitAnimationDuration: 100,
 } as const;