+++ libs/components/src/lib/components/overlay/components/bottom-sheet/partials/bottom-sheet-drag-handle/bottom-sheet-drag-handle.component.ts
@@ -1,16 +1,18 @@
 import {
-  OnInit,
-  OnChanges,
-  Input,
-  HostBinding,
-  ElementRef,
-  SimpleChanges,
-  HostListener,
-  Component,
-  ViewEncapsulation,
   ChangeDetectionStrategy,
+  Component,
+  ElementRef,
+  HostBinding,
+  HostListener,
   inject,
+  Input,
+  OnChanges,
+  OnInit,
+  SimpleChanges,
+  ViewEncapsulation,
 } from '@angular/core';
+import { DestroyService } from '@ethlete/core';
+import { fromEvent, takeUntil, tap } from 'rxjs';
 import { BottomSheetService, BottomSheetSwipeHandlerService } from '../../services';
 import { BottomSheetRef, getClosestBottomSheet } from '../../utils';
 
@@ -26,13 +28,14 @@ import { BottomSheetRef, getClosestBottomSheet } from '../../utils';
   standalone: true,
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
-  providers: [BottomSheetSwipeHandlerService],
+  providers: [BottomSheetSwipeHandlerService, DestroyService],
 })
 export class BottomSheetDragHandleComponent implements OnInit, OnChanges {
   private _bottomSheetRef = inject(BottomSheetRef, { optional: true });
   private readonly _elementRef = inject<ElementRef<HTMLElement>>(ElementRef);
   private readonly _bottomSheetService = inject(BottomSheetService);
   private readonly _bottomSheetSwipeHandlerService = inject(BottomSheetSwipeHandlerService);
+  private readonly _destroy$ = inject(DestroyService, { host: true }).destroy$;
 
   @Input('aria-label')
   ariaLabel?: string = 'Close sheet';
@@ -59,6 +62,27 @@ export class BottomSheetDragHandleComponent implements OnInit, OnChanges {
 
       this._bottomSheetRef = closestRef;
     }
+
+    fromEvent<TouchEvent>(this._elementRef.nativeElement, 'touchstart', { passive: true })
+      .pipe(
+        tap((event) => this._onTouchStart(event)),
+        takeUntil(this._destroy$),
+      )
+      .subscribe();
+
+    fromEvent<TouchEvent>(this._elementRef.nativeElement, 'touchmove', { passive: true })
+      .pipe(
+        tap((event) => this._onTouchMove(event)),
+        takeUntil(this._destroy$),
+      )
+      .subscribe();
+
+    fromEvent<TouchEvent>(this._elementRef.nativeElement, 'touchend', { passive: true })
+      .pipe(
+        tap(() => this._onTouchEnd()),
+        takeUntil(this._destroy$),
+      )
+      .subscribe();
   }
 
   ngOnChanges(changes: SimpleChanges) {
@@ -82,7 +106,6 @@ export class BottomSheetDragHandleComponent implements OnInit, OnChanges {
     );
   }
 
-  @HostListener('touchstart', ['$event'])
   _onTouchStart(event: TouchEvent) {
     if (!this._bottomSheetRef) {
       return;
@@ -94,7 +117,6 @@ export class BottomSheetDragHandleComponent implements OnInit, OnChanges {
     );
   }
 
-  @HostListener('touchmove', ['$event'])
   _onTouchMove(event: TouchEvent) {
     if (!this._swipeHandlerId) {
       return;
@@ -107,7 +129,6 @@ export class BottomSheetDragHandleComponent implements OnInit, OnChanges {
     }
   }
 
-  @HostListener('touchend')
   _onTouchEnd() {
     if (!this._swipeHandlerId) {
       return;