+++ libs/components/src/lib/components/overlay/components/bottom-sheet/stories/components/bottom-sheet-host.storybook.component.ts
@@ -46,12 +46,6 @@ export class BottomSheetHostStorybookComponent {
   disableClose = this._defaultConfig.disableClose;
 
   @Input()
-  enterAnimationDuration = this._defaultConfig.enterAnimationDuration;
-
-  @Input()
-  exitAnimationDuration = this._defaultConfig.exitAnimationDuration;
-
-  @Input()
   hasBackdrop = this._defaultConfig.hasBackdrop;
 
   @Input()