+++ libs/components/src/lib/components/overlay/components/dialog/components/dialog-container/dialog-container.component.ts
@@ -8,11 +8,11 @@ import {
   ElementRef,
   Inject,
   NgZone,
-  OnDestroy,
   Optional,
   ViewEncapsulation,
 } from '@angular/core';
-import { DIALOG_ANIMATION_CLASSES, DIALOG_CONFIG, DIALOG_TRANSITION_DURATION_PROPERTY } from '../../constants';
+import { AnimatedLifecycleDirective, nextFrame } from '@ethlete/core';
+import { DIALOG_CONFIG } from '../../constants';
 import { DialogContainerBaseComponent } from '../../partials';
 import { DialogConfig } from '../../types';
 
@@ -36,13 +36,9 @@ import { DialogConfig } from '../../types';
   },
   standalone: true,
   imports: [PortalModule],
+  hostDirectives: [AnimatedLifecycleDirective],
 })
-export class DialogContainerComponent extends DialogContainerBaseComponent implements OnDestroy {
-  private _hostElement: HTMLElement = this._elementRef.nativeElement;
-  private _openAnimationDuration = this._config.enterAnimationDuration ?? 300;
-  private _closeAnimationDuration = this._config.exitAnimationDuration ?? 100;
-  private _animationTimer: number | null = null;
-
+export class DialogContainerComponent extends DialogContainerBaseComponent {
   constructor(
     elementRef: ElementRef,
     focusTrapFactory: FocusTrapFactory,
@@ -59,62 +55,9 @@ export class DialogContainerComponent extends DialogContainerBaseComponent imple
 
   protected override _contentAttached(): void {
     super._contentAttached();
-    this._startOpenAnimation();
-  }
 
-  override ngOnDestroy() {
-    super.ngOnDestroy();
-
-    if (this._animationTimer !== null) {
-      clearTimeout(this._animationTimer);
-    }
-  }
-
-  private _startOpenAnimation() {
-    setTimeout(() => {
-      this._animationStateChanged.emit({ state: 'opening', totalTime: this._openAnimationDuration });
-
-      this._hostElement.style.setProperty(DIALOG_TRANSITION_DURATION_PROPERTY, `${this._openAnimationDuration}ms`);
-      this._hostElement.classList.add(DIALOG_ANIMATION_CLASSES.opening);
-
-      this._waitForAnimationToComplete(this._openAnimationDuration, this._finishDialogOpen);
+    nextFrame(() => {
+      this._animatedLifecycle.enter();
     });
-  }
-
-  _startExitAnimation(): void {
-    if (this._animationTimer !== null) {
-      clearTimeout(this._animationTimer);
-      this._clearAnimationClasses();
-    }
-
-    this._animationStateChanged.emit({ state: 'closing', totalTime: this._closeAnimationDuration });
-    this._hostElement.classList.remove(DIALOG_ANIMATION_CLASSES.open);
-    this._hostElement.style.setProperty(DIALOG_TRANSITION_DURATION_PROPERTY, `${this._closeAnimationDuration}ms`);
-    this._hostElement.classList.add(DIALOG_ANIMATION_CLASSES.closing);
-    this._waitForAnimationToComplete(this._closeAnimationDuration, this._finishDialogClose);
-  }
-
-  private _finishDialogOpen = () => {
-    this._clearAnimationClasses();
-    this._openAnimationDone(this._openAnimationDuration);
-    this._hostElement.classList.add(DIALOG_ANIMATION_CLASSES.open);
-  };
-
-  private _clearAnimationClasses() {
-    this._hostElement.classList.remove(DIALOG_ANIMATION_CLASSES.opening);
-    this._hostElement.classList.remove(DIALOG_ANIMATION_CLASSES.closing);
-  }
-
-  private _finishDialogClose = () => {
-    this._clearAnimationClasses();
-    this._animationStateChanged.emit({ state: 'closed', totalTime: this._closeAnimationDuration });
-  };
-
-  private _waitForAnimationToComplete(duration: number, callback: () => void) {
-    if (this._animationTimer !== null) {
-      clearTimeout(this._animationTimer);
-    }
-
-    this._animationTimer = window.setTimeout(callback, duration);
   }
 }