+++ libs/components/src/lib/components/overlay/components/dialog/stories/dialog.stories.ts
@@ -65,16 +65,6 @@ export default {
         type: 'boolean',
       },
     },
-    enterAnimationDuration: {
-      control: {
-        type: 'number',
-      },
-    },
-    exitAnimationDuration: {
-      control: {
-        type: 'number',
-      },
-    },
     hasBackdrop: {
       control: {
         type: 'boolean',