+++ libs/components/src/lib/components/overlay/components/toggletip/stories/components/toggletip-storybook.component.ts
@@ -1,15 +1,17 @@
 import { OverlayModule } from '@angular/cdk/overlay';
 import { ChangeDetectionStrategy, Component, ViewEncapsulation } from '@angular/core';
+import { Placement } from '@popperjs/core';
 import { ToggletipDirective } from '../../directives';
 import { ToggletipCloseDirective } from '../../partials';
 
 @Component({
-  selector: 'et-sb-pagination',
+  selector: 'et-sb-toggletip',
   template: `
     <div class="scroll-container">
       <div>
         <button
           [showToggletip]="showTooltip"
+          [placement]="placement"
           (click)="showTooltip = !showTooltip"
           (toggletipClose)="showTooltip = false"
           class="toggletip-trigger"
@@ -26,6 +28,7 @@ import { ToggletipCloseDirective } from '../../partials';
         <button
           [showToggletip]="showTooltip2"
           [etToggletip]="toggletipTpl"
+          [placement]="placement"
           (click)="showTooltip2 = !showTooltip2"
           (toggletipClose)="showTooltip2 = false"
           class="toggletip-trigger"
@@ -70,4 +73,11 @@ import { ToggletipCloseDirective } from '../../partials';
 export class ToggletipStorybookComponent {
   showTooltip = true;
   showTooltip2 = false;
+
+  placement: Placement = 'top';
+
+  set showToggletip(value: boolean) {
+    this.showTooltip = value;
+    this.showTooltip2 = value;
+  }
 }