+++ libs/components/src/lib/components/overlay/components/toggletip/stories/toggletip.stories.ts
@@ -1,3 +1,4 @@
+import { placements } from '@popperjs/core';
 import { Meta, Story } from '@storybook/angular';
 import { ToggletipStorybookComponent } from './components';
 import CustomMDXDocumentation from './toggletip.docs.mdx';
@@ -9,6 +10,22 @@ export default {
     docs: {
       page: CustomMDXDocumentation,
     },
+  },
+  argTypes: {
+    placement: {
+      control: {
+        type: 'select',
+      },
+      options: placements,
+    },
+    showToggletip: {
+      control: {
+        type: 'boolean',
+      },
+    },
+  },
+  args: {
+    placement: 'auto',
   },
 } as Meta<ToggletipStorybookComponent>;